+++ sdk/src/decode/user.ts
@@ -338,7 +338,12 @@ export function decodeUser(buffer: Buffer): UserAccount {
 	}
 	offset += 1;
 
-	// @ts-ignore
+	const poolId = buffer.readUint8(offset);
+	offset += 1;
+	offset += 3; // padding
+
+	const lastFuelBonusUpdateTs = buffer.readUint32LE(offset);
+	offset += 4;
 	return {
 		authority,
 		delegate,
@@ -367,5 +372,7 @@ export function decodeUser(buffer: Buffer): UserAccount {
 		openAuctions,
 		hasOpenAuction,
 		marginMode,
+		poolId,
+		lastFuelBonusUpdateTs,
 	};
 }