+++ sdk/src/util/digest.ts
@@ -11,6 +11,6 @@ export function digestSignature(signature: Uint8Array): string {
 	return createHash('sha256').update(signature).digest('base64');
 }
 
-export function generateSwiftUuid(): Uint8Array {
+export function generateSignedMsgUuid(): Uint8Array {
 	return Uint8Array.from(Buffer.from(nanoid(8)));
 }