+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
+- program: add transfer_pools ix ([#1472](https://github.com/drift-labs/protocol-v2/pull/1472))
+
 ### Fixes
 
 ### Breaking